src/main/java/com/recipeBook/recipebook/converters/UnitOfMeasureCommandToUnitOfMeasure.java
package com.recipeBook.recipebook.converters;


import com.recipeBook.recipebook.commands.UnitOfMeasureCommand;
import com.recipeBook.recipebook.domain.UnitOfMeasure;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

@Component
public class UnitOfMeasureCommandToUnitOfMeasure implements Converter<UnitOfMeasureCommand, UnitOfMeasure> {

    @Synchronized /** thread safe */
    @Nullable
    @Override
    public UnitOfMeasure convert(UnitOfMeasureCommand source) {
        if (source == null) {
            return null;
        }

        final UnitOfMeasure uom = new UnitOfMeasure(); /** immutable */
        uom.setId(source.getId());
        uom.setDescription(source.getDescription());
        return uom;
    }
}
